refactor: 타입 통합 및 import 수정
