+++ src/app/account/components/forms/SignUpForm.tsx
@@ -1,12 +1,12 @@
 import { UseFormReturn } from 'react-hook-form';
-import Input from '../inputs/Input';
-import BaseButton from '../buttons/BaseButton';
-import KakaoLoginButton from '../buttons/KakaoLoginButton';
-import { useKakaoLogin, useLogin, useSignUp } from '@/src/hooks/auth/useAuth';
-import { InitialFormValues, SignUpFormValues } from '../modals/AuthModal';
+import Input from '@/src/components/common/inputs/Input';
+import BaseButton from '@/src/components/common/buttons/BaseButton';
+import KakaoLoginButton from '@/src/components/common/buttons/KakaoLoginButton';
+import { useKakaoLogin, useSignUp } from '@/src/hooks/auth/useAuth';
+import { SignUpFormValues } from '@/src/components/common/modals/AuthModal';
 import { usePathname, useRouter } from 'next/navigation';
 import { useModal } from '@/src/hooks/useModal';
-import Checkbox from '../buttons/CheckBox';
+import Checkbox from '@/src/components/common/buttons/CheckBox';
 
 interface SignUpFormProps {
   signUpForm: UseFormReturn<SignUpFormValues>;
