--- src/types/form.types.ts
@@ -1,7 +1,0 @@
-export interface CheckboxProps {
-  id: string;
-  checked: boolean;
-  onChange: (checked: boolean) => void;
-  label: string;
-  error?: string;
-}
